[#31] Rename lexical analyzer to lexer
